requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Werkzeug==2.3.6
pytz==2023.3
PyMySQL==1.1.0
cryptography==41.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
